Retire naming case-rules registry seam
